Feat: Move ProjectPreview component to preview (#19)
