+++ src/app/[locale]/projects/create/page.tsx
@@ -9,7 +9,7 @@ import { useTranslations } from 'next-intl';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import * as z from 'zod';
-import { ProjectPreview } from '@/components/cards/ProjectPreview';
+import { ProjectPreview } from '@/components/previews/ProjectPreview';
 import { Button } from '@/components/ui/button';
 import { Calendar } from '@/components/ui/calendar';
 import {
